+++ HISTORY.rst
@@ -3,6 +3,13 @@
 History
 -------
 
+1.8.1 (2014-05-04)
+~~~~~~~~~~~~~~~~~~
+
+- Changing the Gazibit report for shared resources to include the model
+  shared url in embedded format.
+- Fixing bug: train and tests data could not be read from stdin
+
 1.8.0 (2014-04-29)
 ~~~~~~~~~~~~~~~~~~